+++ supertokens/main.go
@@ -24,7 +24,7 @@ func Init(config TypeInput) error {
 }
 
 func Middleware(theirHandler http.Handler) http.Handler {
-	instance, err := getInstanceOrThrowError()
+	instance, err := GetInstanceOrThrowError()
 	if err != nil {
 		panic("Please call supertokens.Init function before using the Middleware")
 	}
@@ -32,7 +32,7 @@ func Middleware(theirHandler http.Handler) http.Handler {
 }
 
 func ErrorHandler(err error, req *http.Request, res http.ResponseWriter) error {
-	instance, instanceErr := getInstanceOrThrowError()
+	instance, instanceErr := GetInstanceOrThrowError()
 	if instanceErr != nil {
 		return instanceErr
 	}
@@ -40,7 +40,7 @@ func ErrorHandler(err error, req *http.Request, res http.ResponseWriter) error {
 }
 
 func GetAllCORSHeaders() []string {
-	instance, err := getInstanceOrThrowError()
+	instance, err := GetInstanceOrThrowError()
 	if err != nil {
 		panic("Please call supertokens.Init before using the GetAllCORSHeaders function")
 	}